sin terminar creo

+++ docker-entrypoint.sh
@@ -1,7 +1,9 @@
-#!/bin/bash
+#!/bin/sh
 set -e
 
-cp /app/.env.docker /app/.env
+if [ ! -f /app/.env ] && [ -f /app/.env.docker ]; then
+	cp /app/.env.docker /app/.env
+fi
 php artisan key:generate --force
 
 echo "Notificaciones listo."
